Moves MustacheTemplateRenderer to the Services folder
